Adding Spanish, French and German languages
